+++ src/main/java/com/labSync/LabSync/controller/ProtocolMethods.java
@@ -1,6 +1,5 @@
 package com.labSync.LabSync.controller;
 
-import com.labSync.LabSync.models.User;
 import org.springframework.http.ResponseEntity;
 
 import java.util.List;
@@ -15,7 +14,7 @@ public interface ProtocolMethods<T> {
 
     ResponseEntity<T> patch(long id, T entity);
 
-    ResponseEntity<T> getById(int id);
+    ResponseEntity<T> getById(long id);
 
     ResponseEntity<List<T>> getAll();
